Move widget signals into wirecloud.platform.signals.handlers
